refactor: move Ids to upload with tests
